Add support for office hours
Addding functions to count number of worked hours between two times,
as well as adding some work time to a given time.

Work time considers all days of the week are worked the same hours and
all holidays are full-day holidays.

timezones and daylight savings are left to the responsibility of the go
time implementation. Hence any trouble in this area should be reported
to go