v2/au: add ANZAC rules for different states
In NSW/VIC no day in lou for ANZAC day is observed.

In NT/QLD/SA if ANZAC day is on a Sunday the following Monday is
observed as a public holiday.

In ACT/WA if ANZAC day falls on a weekend the following Monday is
observed as a public holiday.

This commit adds these rules to correctly capture ANZAC day holidays.

The best way to see this is looking at 2020 (ANZAC day is on Sat) and
2021 (ANZAC day is on a Sunday).

https://www.fairwork.gov.au/leave/public-holidays/list-of-public-holidays